refactor: use URL instead of str for representing URLs, don't hard-code the name of the xml file
