+++ tests/test_cli.py
@@ -4,10 +4,12 @@ import os
 from click.testing import CliRunner
 from matchviz.cli import tabulate_matches_cli
 
+
+@pytest.mark.skip
 @pytest.mark.parametrize(
     "alignment_url",
     [
-        "s3://aind-open-data/exaSPIM_708373_2024-04-02_19-49-38_alignment_2024-05-07_18-15-25/"
+        "s3://aind-open-data/exaSPIM_708373_2024-04-02_19-49-38_alignment_2024-05-07_18-15-25/bigstitcher.xml"
     ],
 )
 def test_save_points(tmpdir, alignment_url):
@@ -24,11 +26,12 @@ def test_save_points(tmpdir, alignment_url):
     )
     assert run_result.returncode == 0
 
+
 @pytest.mark.skip
 @pytest.mark.parametrize(
     "alignment_url",
     [
-        "s3://aind-open-data/exaSPIM_708373_2024-04-02_19-49-38_alignment_2024-05-07_18-15-25/"
+        "s3://aind-open-data/exaSPIM_708373_2024-04-02_19-49-38_alignment_2024-05-07_18-15-25/bigstitcher.xml"
     ],
 )
 def test_save_neuroglancer(tmpdir, alignment_url):
@@ -48,8 +51,9 @@ def test_save_neuroglancer(tmpdir, alignment_url):
     )
     assert run_result.returncode == 0
 
+
 @pytest.mark.skip
 def test_summarize_points():
     alignment_url = "s3://aind-open-data/exaSPIM_708373_2024-04-02_19-49-38_alignment_2024-05-07_18-15-25/"
     runner = CliRunner()
-    cli_result = runner.invoke(tabulate_matches_cli, [f"--alignment-url", alignment_url])
+    _ = runner.invoke(tabulate_matches_cli, ["--alignment-url", alignment_url])
